moved DatasetError and DatasetSample into types.rs
