+++ examples/common/datasets/mod.rs
@@ -2,8 +2,10 @@ pub mod euroc;
 pub mod kitti;
 pub mod rectify;
 pub mod stereo_calib;
+pub mod types;
 
 pub use euroc::EurocDataset;
 pub use kitti::KittiDataset;
 pub use rectify::StereoRectifier;
 pub use stereo_calib::StereoCalib;
+pub use types::{DatasetError, DatasetSample};
